refactor(tags): create protocol for how tags are arrange in both views
