+++ PocketKit/Tests/PocketKitTests/PocketAddTagsViewModelTests.swift
@@ -1,6 +1,7 @@
 import XCTest
 import Combine
 import Analytics
+import Textile
 
 @testable import Sync
 @testable import PocketKit
@@ -43,7 +44,7 @@ class PocketAddTagsViewModelTests: XCTestCase {
         }
 
         let viewModel = subject(item: item) { }
-        let isValidTag = viewModel.addTag(with: "tag 2")
+        let isValidTag = viewModel.addNewTag(with: "tag 2")
 
         XCTAssertTrue(isValidTag)
         XCTAssertEqual(viewModel.tags, ["tag 1", "tag 2"])
@@ -56,7 +57,7 @@ class PocketAddTagsViewModelTests: XCTestCase {
         }
 
         let viewModel = subject(item: item) { }
-        let isValidTag = viewModel.addTag(with: "tag 1")
+        let isValidTag = viewModel.addNewTag(with: "tag 1")
 
         XCTAssertFalse(isValidTag)
         XCTAssertEqual(viewModel.tags, ["tag 1"])
@@ -69,7 +70,7 @@ class PocketAddTagsViewModelTests: XCTestCase {
         }
 
         let viewModel = subject(item: item) { }
-        let isValidTag = viewModel.addTag(with: "  ")
+        let isValidTag = viewModel.addNewTag(with: "  ")
 
         XCTAssertFalse(isValidTag)
         XCTAssertEqual(viewModel.tags, ["tag 1"])
@@ -116,7 +117,7 @@ class PocketAddTagsViewModelTests: XCTestCase {
         viewModel.allOtherTags()
 
         wait(for: [expectRetrieveTagsCall], timeout: 1)
-        XCTAssertEqual(viewModel.otherTags, ["tag 2", "tag 3"])
+        XCTAssertEqual(viewModel.otherTags, [TagType.tag("tag 3"), TagType.tag("tag 2")])
         XCTAssertNotNil(source.retrieveTagsCall(at: 0))
     }
 
@@ -130,7 +131,7 @@ class PocketAddTagsViewModelTests: XCTestCase {
         viewModel.removeTag(with: "tag 2")
 
         XCTAssertEqual(viewModel.tags, ["tag 1", "tag 3"])
-        XCTAssertEqual(viewModel.otherTags, ["tag 2"])
+        XCTAssertEqual(viewModel.otherTags, [TagType.tag("tag 2")])
     }
 
     func test_removeTag_withNotExistingName_updatesTags() {
@@ -168,7 +169,7 @@ class PocketAddTagsViewModelTests: XCTestCase {
             .delay(for: .seconds(1), scheduler: RunLoop.main, options: .none)
             .sink { string in
                 defer { expectFilterTagsCall.fulfill() }
-                XCTAssertEqual(viewModel.otherTags, ["tag 2", "tag 3"])
+                XCTAssertEqual(viewModel.otherTags, [TagType.tag("tag 2"), TagType.tag("tag 3")])
             }
             .store(in: &subscriptions)
         wait(for: [expectFilterTagsCall], timeout: 1)
@@ -197,7 +198,7 @@ class PocketAddTagsViewModelTests: XCTestCase {
             .delay(for: .seconds(1), scheduler: RunLoop.main, options: .none)
             .sink { string in
                 defer { expectFilterTagsCall.fulfill() }
-                XCTAssertEqual(viewModel.otherTags, ["tag 2", "tag 3"])
+                XCTAssertEqual(viewModel.otherTags, [TagType.tag("tag 2"), TagType.tag("tag 3")])
             }
             .store(in: &subscriptions)
         wait(for: [expectFilterTagsCall], timeout: 1)
